refactor sample Responsive from sandbox
