+++ sample/Responsive/Controllers/HomeController.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2019 Sarin Na Wangkanai, All Rights Reserved.
+// Copyright (c) 2019 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
 using System;
@@ -10,7 +10,7 @@ using Microsoft.AspNetCore.Mvc;
 using Wangkanai.Detection;
 using Wangkanai.Responsive;
 
-namespace Sandbox.Controllers
+namespace Responsive.Controllers
 {
     public class HomeController : Controller
     {
