+++ sample/Responsive/Views/Home/Index.cshtml
@@ -1,4 +1,6 @@
-﻿
+
 <p>Views/Home/Index.cshtml</p>
 
 @Html.Partial("_TagHelper")
+
+<
