+++ sample/Responsive/Views/Shared/_TagHelper.cshtml
@@ -1,4 +1,4 @@
-﻿<div class="taghelper">
+<div class="taghelper">
     <h4>desktop</h4>
     <device include="desktop">is desktop</device>
     <device exclude="desktop">not desktop</device>
