+++ sample/Responsive/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-﻿@using Sandbox
+@using Responsive
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Wangkanai.Responsive
